Refactor AnnoX classes MatchSpan class is now more generic SpanGroup AnnoX classes have separate files AnnoTier without_overlaps method added and unused methods removed
